src/data_collection/knowyourmeme_scraper.py
import requests
from bs4 import BeautifulSoup
import pandas as pd
from config.settings import KNOWYOURMEME_URL

class KnowYourMemeCollector:
    def __init__(self):
        self.base_url = KNOWYOURMEME_URL
        
    def fetch_weaponized_patterns(self):
        """Identify weaponized meme patterns for brand safety"""
        
        weaponized_patterns = [
            {'meme_name': 'pump and dump', 'toxicity_level': 'high', 'financial_risk': 0.9},
            {'meme_name': 'rug pull', 'toxicity_level': 'critical', 'financial_risk': 0.95},
            {'meme_name': 'coordinated raid', 'toxicity_level': 'high', 'financial_risk': 0.85},
            {'meme_name': 'fake news campaign', 'toxicity_level': 'critical', 'financial_risk': 0.92},
            {'meme_name': 'manipulation squad', 'toxicity_level': 'high', 'financial_risk': 0.88}
        ]
        
        df = pd.DataFrame(weaponized_patterns)
        print(f"Weaponized patterns identified: {len(df)}")
        
        return df
    
    def search_meme_origin(self, meme_term):
        """Search for meme origin and context"""
        
        try:
            search_url = f"{self.base_url}/search?q={meme_term}"
            response = requests.get(search_url, timeout=10)
            
            if response.status_code == 200:
                soup = BeautifulSoup(response.content, 'html.parser')
                return {
                    'term': meme_term,
                    'found': True,
                    'origin_data': 'Available'
                }
        except Exception as e:
            print(f"Error searching {meme_term}: {e}")
        
        return {'term': meme_term, 'found': False, 'origin_data': None}

if __name__ == "__main__":
    collector = KnowYourMemeCollector()
    patterns = collector.fetch_weaponized_patterns()
    print(patterns)
