requirements.txt
google-cloud-bigquery>=3.11.0
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
vertexai>=1.38.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
plotly>=5.17.0
pytrends>=4.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
seaborn>=0.12.0
matplotlib>=3.7.0
python-dateutil>=2.8.0